question-generation-app/src/app/shared/ui-components/copy-to-clipboard/copy-to-clipboard.component.ts
import { Component, Input, OnInit } from '@angular/core';

@Component({
  selector: 'app-copy-to-clipboard',
  templateUrl: './copy-to-clipboard.component.html',
  styleUrls: ['./copy-to-clipboard.component.scss']
})
export class CopyToClipboardComponent implements OnInit {
  @Input()
  itemToCopy: String;

  @Input()
  iconClass: String;

  constructor() {}

  ngOnInit() {}

  copyId() {
    const selBox = document.createElement('textarea');
    selBox.style.position = 'fixed';
    selBox.style.left = '0';
    selBox.style.top = '0';
    selBox.style.opacity = '0';
    selBox.value = String(this.itemToCopy);
    document.body.appendChild(selBox);
    selBox.focus();
    selBox.select();
    document.execCommand('copy');
    document.body.removeChild(selBox);
  }
}
